+++ spec/model/ios_spec.rb
@@ -45,4 +45,25 @@ describe 'model/IOS' do
     _(result.to_cfg).wont_match(/public/)
     _(result.to_cfg).wont_match(/AAAAAAAAAABBBBBBBBBB/)
   end
+
+  it 'removes secrets from IOS-XE WLCs' do
+    Oxidized.config.vars.remove_secret = true
+    mockmodel = MockSsh.new('examples/device-simulation/yaml/iosxe_C9800-L-F-K9_17.06.05.yaml')
+    Net::SSH.stubs(:start).returns mockmodel
+
+    status, result = @node.run
+
+    _(status).must_equal :success
+    _(result.to_cfg).wont_match(/SECRET_REMOVED/)
+    _(result.to_cfg).wont_match(/REMOVED_SECRET/)
+    _(result.to_cfg).wont_match(/WLANSECR3T/)
+    _(result.to_cfg).wont_match(/WLAN SECR3T/)
+    _(result.to_cfg).wont_match(/7df35f90c92ecff2a803e79577b85e978edc0a76404f6cfb534df8d9f9f67beb/)
+    _(result.to_cfg).wont_match(/DOT1XPASSW0RD/)
+    _(result.to_cfg).wont_match(/MGMTPASSW0RD/)
+    _(result.to_cfg).wont_match(/MGMTSECR3T/)
+    _(result.to_cfg).wont_match(/DOT1X PASSW0RD/)
+    _(result.to_cfg).wont_match(/MGMT PASSW0RD/)
+    _(result.to_cfg).wont_match(/MGMT SECR3T/)
+  end
 end